Remove KickstarterUITests target
